refactor(backend): move graph DSL sanitizer policy into whiteboard domain (#505)
