+++ backend/app/domain/whiteboard/__init__.py
@@ -1,0 +1,5 @@
+"""Whiteboard domain package."""
+
+from .graph_dsl import sanitize_whiteboard_dsl
+
+__all__ = ["sanitize_whiteboard_dsl"]
